+++ server/mediaserver.go
@@ -28,7 +28,7 @@ import (
 	"time"
 
 	"github.com/livepeer/go-livepeer/clog"
-	"github.com/livepeer/go-livepeer/mediaserver"
+	"github.com/livepeer/go-livepeer/media"
 	"github.com/livepeer/go-livepeer/monitor"
 	"github.com/livepeer/go-livepeer/pm"
 	"github.com/livepeer/go-tools/drivers"
@@ -127,7 +127,7 @@ type LivepeerServer struct {
 	connectionLock    *sync.RWMutex
 	serverLock        *sync.RWMutex
 
-	mediaMTXClient   *mediaserver.MediaMTXClient
+	mediaMTXClient   *media.MediaMTXClient
 	liveAIAuthApiKey string
 }
 
@@ -194,7 +194,7 @@ func NewLivepeerServer(rtmpAddr string, lpNode *core.LivepeerNode, httpIngest bo
 		internalManifests:       make(map[core.ManifestID]core.ManifestID),
 		recordingsAuthResponses: cache.New(time.Hour, 2*time.Hour),
 		AISessionManager:        NewAISessionManager(lpNode, AISessionManagerTTL),
-		mediaMTXClient:          mediaserver.NewMediaMTXClient(lpNode.MediaMTXApiPassword),
+		mediaMTXClient:          media.NewMediaMTXClient(lpNode.MediaMTXApiPassword),
 		liveAIAuthApiKey:        lpNode.LiveAIAuthApiKey,
 	}
 	if lpNode.NodeType == core.BroadcasterNode && httpIngest {
